Gbrowse-track like SVG with read phases
It's pretty fragile at the moment (have to manually hard-code the
position of the gene you want to look at), but it does seem to work.
Since I figure reads without any SNPs are uninteresting, I throw them
out (but I do plot the mates of ones that have SNPs).  There does seem
to be an issue with the first splice junction on reads in the forward
strand.  Will work that out soon.