1st stab at figuring out how good are replicates
For each slice, find the nearest slice to it in each other embryo, then
find the correlation for ase and expression between the values in that
slice.  One would expect that an embryo will be more closely related to
it's replicates than those of other genotypes. This does seem to be the
case! Yay!