Subtract off the mean of the ASE values to look at shape
One potential argument for this approach is that, if there's a
systematic mapping bias, then the absolute FPKMs from each of the two
species won't be directly comparable.  Subtracting off the difference in
means is a quick and (very) dirty way to compensate for that.

Other approaches could be:

 a) Use WASP.  In principle, rather than generating the complete listing
of all possible substitutions RRR, RRA, RAR, RAA, ARR, ARA, AAR, AAA
(for a read with 3 snps, each of which is ref/alt), I only need to
generate RRR and AAA, since this is a hybrid, rather than a homogenous
population.  This should make it hugely faster.

 b) For each gene with a 1-to-1 ortholog in simulans, see if I can
estimate the conversion factor between the FPKM in simulans embryos
mapped on a simulans genome (which is presumably the "real" expression
level) and the simulans expression level mapped on the mel genome.  This
has the disadvantage of being hacky, hard to explain, and only liable to
work for some genes.