Be more explicit in which stage I'm using
This fixes a bug, where I thought that the second-to-last stage was the
same, whereas in mel it was the right stage, but sim has a stage 6a
timepoint.  Switching to the correct stage, as well as normalizing based
on the 90th percentile expression only in the posterior stripe, rather
than 90th percentile of all except the inter-stripe.