Bugfix for index keyword
Previously, if they index was from a series that had non-contiguous
numbering (perhaps due to sorting something else that was numbered, and
only taking a subset), then indexing it later on would have problems.
This way, it converts to a list, so using square brackets with a
numbered index does the Right Thing (tm).