Option to have multiple, aligned labels
This is useful for outputting not just gene names, but also R^2 values,
FPKM summaries, chromosomes, etc. Instead of a list of strings, you pass
in a list of tuples of strings, and it will create a group of text
elements. I just hardcoded the width, so some tweaking in Illustrator
may be necessary afterwards.